Factor Kore.Step.Proof out of Kore.Step.Step
